+++ app/configurator/components/chart-options-selector.tsx
@@ -576,7 +576,7 @@ const EncodingOptionsPanel = (props: EncodingOptionsPanelProps) => {
                   const { checked } = e.target;
                   if ("y" in fields && !("showDots" in fields.y)) {
                     dispatch({
-                      type: "COLOR_MAPPING_UPDATED",
+                      type: "COLOR_FIELD_UPDATED",
                       value: {
                         locale,
                         field: "y",
@@ -586,7 +586,7 @@ const EncodingOptionsPanel = (props: EncodingOptionsPanelProps) => {
                     });
                   }
                   dispatch({
-                    type: "COLOR_MAPPING_UPDATED",
+                    type: "COLOR_FIELD_UPDATED",
                     value: {
                       locale,
                       field: encoding.field,
@@ -1194,7 +1194,7 @@ const ChartComboLineSingleYField = ({
                   }
 
                   dispatch({
-                    type: "COLOR_MAPPING_UPDATED",
+                    type: "COLOR_FIELD_UPDATED",
                     value: {
                       locale,
                       field: "y",
@@ -1222,7 +1222,7 @@ const ChartComboLineSingleYField = ({
 
                 if (id !== FIELD_VALUE_NONE) {
                   dispatch({
-                    type: "COLOR_MAPPING_UPDATED",
+                    type: "COLOR_FIELD_UPDATED",
                     value: {
                       locale,
                       field: "y",
@@ -1311,7 +1311,7 @@ const ChartComboLineDualYField = ({
             onChange={(e) => {
               const newId = e.target.value as string;
               dispatch({
-                type: "COLOR_MAPPING_UPDATED",
+                type: "COLOR_FIELD_UPDATED",
                 value: {
                   locale,
                   field: "y",
@@ -1336,7 +1336,7 @@ const ChartComboLineDualYField = ({
             onChange={(e) => {
               const newId = e.target.value as string;
               dispatch({
-                type: "COLOR_MAPPING_UPDATED",
+                type: "COLOR_FIELD_UPDATED",
                 value: {
                   locale,
                   field: "y",
@@ -1426,7 +1426,7 @@ const ChartComboLineColumnYField = ({
             onChange={(e) => {
               const newId = e.target.value as string;
               dispatch({
-                type: "COLOR_MAPPING_UPDATED",
+                type: "COLOR_FIELD_UPDATED",
                 value: {
                   locale,
                   field: "y",
@@ -1451,7 +1451,7 @@ const ChartComboLineColumnYField = ({
             onChange={(e) => {
               const newId = e.target.value as string;
               dispatch({
-                type: "COLOR_MAPPING_UPDATED",
+                type: "COLOR_FIELD_UPDATED",
                 value: {
                   locale,
                   field: "y",
@@ -1849,7 +1849,7 @@ const ChartFieldSorting = ({
   >(
     ({ sortingType, sortingOrder }) => {
       dispatch({
-        type: "COLOR_MAPPING_UPDATED",
+        type: "COLOR_FIELD_UPDATED",
         value: {
           locale,
           field,
@@ -2165,11 +2165,7 @@ const ChartFieldColorComponent = ({
           ) : null
         ) : colorType === "numerical" ? (
           <div>
-            <ColorRampField
-              field={field}
-              path="color.paletteId"
-              nSteps={nbClass}
-            />
+            <ColorRampField field={field} path="color" nSteps={nbClass} />
             <FieldSetLegend
               legendTitle={t({
                 id: "controls.scale.type",
@@ -2344,7 +2340,7 @@ const ChartMapBaseLayerSettings = ({
     if (chartConfig.baseLayer.locked) {
       if (map !== null) {
         dispatch({
-          type: "COLOR_MAPPING_UPDATED",
+          type: "COLOR_FIELD_UPDATED",
           value: {
             locale,
             field: null,
@@ -2356,7 +2352,7 @@ const ChartMapBaseLayerSettings = ({
       }
     } else {
       dispatch({
-        type: "COLOR_MAPPING_UPDATED",
+        type: "COLOR_FIELD_UPDATED",
         value: {
           locale,
           field: null,