+++ app/docs/tooltip.stories.tsx
@@ -95,7 +95,7 @@ const TooltipBoxStory = () => (
           chartType: "column",
           fields,
           interactiveFiltersConfig: {
-            legend: { active: false, componentId: "" },
+            legend: { active: false, componentId: "", },
             timeRange: {
               active: false,
               componentId: "",
@@ -261,7 +261,7 @@ const TooltipContentStory = {
           chartType: "column",
           fields,
           interactiveFiltersConfig: {
-            legend: { active: false, componentId: "" },
+            legend: { active: false, componentId: "",  },
             timeRange: {
               active: false,
               componentId: "",
@@ -330,7 +330,7 @@ export const TooltipContentStory2 = {
           chartType: "column",
           fields,
           interactiveFiltersConfig: {
-            legend: { active: false, componentId: "" },
+            legend: { active: false, componentId: "",},
             timeRange: {
               active: false,
               componentId: "",